Change BaseMixin.sort field to BigInteger
To accomodate long sort values caused by
concatenation of parent sort values.
Tests: _calc_sort fail - need to fix how
	after_object is handled if none

+++ tracuse_project/utils/mixins.py
@@ -29,9 +29,9 @@ class BaseMixin(models.Model):
     active = models.BooleanField(default=True,
                                  db_index=True
                                  )
-    sort = models.IntegerField(default=0,
-                               db_index=True
-                               )
+    sort = models.BigIntegerField(default=0,
+                                  db_index=True
+                                  )
     sort_base_zero_fill = 3
     sort_parts = []
